G3 2023 v5

+++ DuggaSys/fileed.js
@@ -84,10 +84,11 @@ function returnedFile(data) {
         extension: "Extension",
         kind: "Kind",
         filesize: "Size",
+        type: "Type",
         uploaddate: "Upload date",
         editor: ""
     }
-    var colOrderPre = ["filename", "extension", "kind", "filesize", "uploaddate", "editor"];
+    var colOrderPre = ["filename", "extension", "kind", "filesize", "type", "uploaddate", "editor"];
 
     if (data['studentteacher'] || data['supervisor']) {
         document.getElementById('fabButton').style.display = "none";
@@ -345,11 +346,14 @@ function renderCell(col, celldata, cellid) {
     if (col == "trashcan" || col == "filename" || col == "filesize" || col == "editor") {
         obj = JSON.parse(celldata);
     }
-
     if (col == "trashcan") {
         if (obj.showtrashcan) {
-            str = "<span class='iconBox'><img alt='delete file icon' tabindex='0' id='dorf' title='Delete file' class='trashcanIcon trashcanTab' src='../Shared/icons/Trashcan.svg' ";
-            str += " onclick='deleteFile(\"" + obj.fileid + "\",\"" + obj.filename + "\",\"" + obj.filekind + "\");' ></span>";
+            if(obj.filePath.includes("Github")) {
+                str = "<span class='iconBox'><img alt='github icon' tabindex='0' id='dorf' title='Github file' style='width:16px' src='../Shared/icons/githubLink-icon.png' ";
+            } else {
+                str = "<span class='iconBox'><img alt='delete file icon' tabindex='0' id='dorf' title='Delete file' class='trashcanIcon trashcanTab' src='../Shared/icons/Trashcan.svg' ";
+                str += " onclick='deleteFile(\"" + obj.fileid + "\",\"" + obj.filename + "\",\"" + obj.filekind + "\");' ></span>";
+            }
         }
     } else if (col == "filename") {
         if (obj.kind == "Link") {
@@ -366,17 +370,25 @@ function renderCell(col, celldata, cellid) {
     } else if (col == "extension" || col == "uploaddate") {
         str += "<span>" + celldata + "</span>";
     } else if (col == "editor") {
-        if(obj.showeditor){
-        if (obj.extension == "md" || obj.extension == "txt" || obj.extension == "html") {
-            str = "<span class='iconBox' ><img alt='edit file icon' tabindex='0' id='dorf'  title='Edit file' class='markdownIcon markdownIconeditFile' src='../Shared/icons/markdownPen.svg' ";
-            str += "onclick='loadPreview(\"" + obj.filePath + "\", \"" + obj.filename + "\", " + obj.kind + ")'></span>";
-        } else if (obj.extension == "js"  || obj.extension == "css" || obj.extension == "php") {
-            str = "<span class='iconBox'><img alt='edit file icon' id='dorf'  title='Edit file'  class='markdownIcon' src='../Shared/icons/markdownPen.svg' ";
-            str += "onclick='loadFile(\"" + obj.filePath + "\", \"" + obj.filename + "\", " + obj.kind + ")'></span>";
+        if (obj.showeditor) {
+            if (!obj.filePath.includes("Github")) {
+                if (obj.extension == "md" || obj.extension == "txt" || obj.extension == "html") {
+                    str = "<span class='iconBox' ><img alt='edit file icon' tabindex='0' id='dorf'  title='Edit file' class='markdownIcon markdownIconeditFile' src='../Shared/icons/markdownPen.svg' ";
+                    str += "onclick='loadPreview(\"" + obj.filePath + "\", \"" + obj.filename + "\", " + obj.kind + ")'></span>";
+                } else if (obj.extension == "js" || obj.extension == "css" || obj.extension == "php") {
+                    str = "<span class='iconBox'><img alt='edit file icon' id='dorf'  title='Edit file'  class='markdownIcon' src='../Shared/icons/markdownPen.svg' ";
+                    str += "onclick='loadFile(\"" + obj.filePath + "\", \"" + obj.filename + "\", " + obj.kind + ")'></span>";
+                }
+            }
         }
-    }
     } else if (col == "kind") {
         str += "<span>" + convertFileKind(celldata) + "</span>";
+    } else if (col == "type") {
+        if(celldata !== null) {
+            str = "<span>" + "Github" + "</span>";
+        } else {
+            str += "<span>" + "Manual" + "</span>";
+        }
     }
     return str;
 }
@@ -485,6 +497,12 @@ function filterFilesByKind(kind){
 
     }if(kind == "Link"){
         $( "td:contains('Link')" ).parents("tr").show();
+    }if(kind == "Manual"){
+        $( "td:contains('Manual')" ).parents("tr").show();
+
+    }if(kind == "Github"){
+        $( "td:contains('Github')" ).parents("tr").show();
+
     }else if(kind == "AllFiles"){
         $("#fileLink table tr").show();
     }
@@ -510,18 +528,18 @@ function sortFiles(asc){
     while(switching){
         switching = false;
         rows = $("#fileLink table tr");
-        for(i = 1; i < (rows.length - 1); i++){
+        for(i = 1; i < rows.length - 1; i++){
             shouldSwitch = false;
             x = rows[i].getElementsByTagName("TD")[3];
             y = rows[i + 1].getElementsByTagName("TD")[3];
 
-           if(asc == true){
+           if(asc == true && y){
            	 if(x.innerHTML.toLowerCase() > y.innerHTML.toLowerCase()){
                 shouldSwitch = true;
                 break;
              }
 
-           }else if(asc == false){
+           }else if(asc == false && y){
            	  if(x.innerHTML.toLowerCase() < y.innerHTML.toLowerCase()){
                 shouldSwitch = true;
                 break;
@@ -677,7 +695,26 @@ function compare(a, b) {
 		}else if(col=="editor"||col=="trashcan"){
 				tempA=-1;
 				tempB=-1;
-		}
+		} else if(col=="type") {
+            if(tempA == null && tempB == null)  {
+                tempA = 0;
+                tempB = 0;
+            } else if (tempA !== null && tempB == null) {
+                tempA = 1;
+                tempB = 0;
+            } else if (tempA == null && tempB !== null) {
+                tempA = 0;
+                tempB = 1;
+            } else {
+                tempA = 1;
+                tempB = 1;
+            }
+/*             if(tempA.path !== null && tempB.path == null) {
+                console.log("tempA är inte null");
+            } else {
+                console.log("tempA är null");
+            } */
+        }
 
     if (tempA > tempB) {
         return 1;
@@ -992,4 +1029,3 @@ document.addEventListener('keydown', function(e) {
 });
 
 
-
